timezone has now moved to a new location
Wrap import in a try/catch to support older versions of airflow,
otherwise import from task sdk
